Delivering HOPS 3.22 vendor branch at Haystack svn rev 3226 from tarball (Mar 22 02:57)/swc/tarpit/hops-tc-3.22swc.tar.gz
This update was driven by chops fixes for VGOS to be captured
in 2.5 and tagged in 2.5.4 (in a few days).  Changes over the
past year include:
    boosted alist extent number 999 -> 9999
    fileset.h  MAXFSET size from 3k to 10k
    vex.h  MAXSCANS from 2048 to 8192
    chops vpal_module module (proxy-cable-cal)
    chops select_bandpols.py



git-svn-id: https://svn.atnf.csiro.au/difx@9983 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/hops/trunk/chops/source/c_src/dfio/src/read_mk4file.c
@@ -24,7 +24,7 @@ int
 read_mk4file (FILE *fp,
               char **ptr)
     {
-    int nb;
+    size_t nb;
     int fd;
     size_t size;
     struct stat file_status;
@@ -57,7 +57,7 @@ read_mk4file (FILE *fp,
                                         /* Did it go OK? */
     if (nb != size)
         {
-        msg ("Error, expected %u bytes, read %d bytes", 2, size, nb);
+        msg ("Error, expected %zu bytes, read %zu bytes", 2, size, nb);
         return (-1);
         }
